Expanded card_decks Module; Created Basic Poker
card_decks: added methods to standard class allowing cards to be dealt
to hand lists (later to be classes themselves) and then removed from the
object deck.

app: Created the basic flow of the poker game, dealing hands, offering
choices, etc. Completion of this and beautificaion are next on my list.